set products.py posgres connection from Flask config
Prior to this change, we were loading the Flask instance config module
directly in products.py.

This makes it hard to run the test suite, because we had to copy
./config.py to ./instance/config.py.

This is complicated by the fact that setuptools does not store the
sample config.py in the "python setup.py sdist" tarball, so it's even
harder to run %check during the rpmbuild.

Move the default config into a "product_listings_manager.config" module.

Instruct the user to override the config settings with the new
FLASK_CONFIG environment variable.

In the product_listings_manager application, after loading the products
module, set the module's database variables from our Flask config's
values.