fix Fault raising
Flask-XML-RPC's "Fault" class is just xmlrpclib.Fault, and that takes
two arguments: a fault code and a fault string. We were only passing the
fault string, so we crashed in the Fault constructor here.

Prior to this change, something in xmlrpclib was casting the exception
to some other data type. For example, if we caught pg.InternalError,
xmlrpclib was translating that to a dict when sending that back through
Flask to the client.

Print a generic string to the client so we don't accidentally leak
information.