setup.py: do not hardcode version number
Prior to this change, we had to update the application version number in
setup.py and in __init__.py.

Update setup.py to read the version number dynamically from __init__.py.

The purpose of this change is to make it easier to bump the version
number value when making releases.